refactor 패키지 이동 (Person, Participants)
